backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.11.1
certifi==2026.1.4
distro==1.9.0
Django==4.2.7
django-cors-headers==4.9.0
djangorestframework==3.14.0
google-generativeai==0.3.0
python-dotenv==1.0.0
pytz==2025.2
sniffio==1.3.1
sqlparse==0.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
